requirements.txt
# Core
pytz==2021.3
Django==3.2.9
django-configurations==2.2
gunicorn==20.1.0
newrelic==7.2.2.169

# For the persistence stores
psycopg2-binary==2.9.1
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.2.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.12.4
Markdown==3.3.4
django-filter==21.1

# Developer Tools
ipdb==0.13.9
mkdocs==1.2.3
flake8==4.0.1

# Testing
mock==4.0.3
factory-boy==3.2.1
django-nose==1.4.7
coverage==6.1.1

# Static and Media Storage
django-storages==1.12.3
boto3==1.19.7

# channels
channels==3.0.4


